Provide support for querying some CPU information with CPUID, improve QEMU configuration yet again and fix a bug
Change QEMU configuration to use the KVM accelerator instead of TCG for higher
performance and the host CPU configuration instead of max which KVM enables
(this allows to for instance confirm that the version and brand
identification features of CPUID work correctly).
    Implement code obtaining some information from CPUID leaves EAX = 0 and
EAX = 1, importing the arrayvec and bitflags crates in the process. Also
implement a trait called AsSliceOf, which defines a function returning a
reference to a fixed-size slice representation of a different element type of
the value passed in as an argument.
    Fix a bug in the Idt struct's Index implementations where they used the
wrong index into the user_defined interrupt descriptor array.