Improve reliability of builds across machines, implement code querying some APIC-related information and add a page fault exception handler
Add a rust-toolchain.toml file that describes the version of Rust nigthly required to build projects in this repository. Modify the rustfmt.toml file to reflect that addition, specifying that version of Rust nightly as the one required to format the code in this repository with Rustfmt. Implement some APIC-related functionality, a part of which is reading a memory mapped LAPIC register, which generates a page fault, for the detection of which a page fault exception handler is created.
    And finally, a README document was created, describing the build process for the kernel.