Discover an obscure bug where the disk_image.bin file gets corrupted every once in a while, preventing the emulated computer from booting up
The corruption happens because files are copied into the image file into increasing physical addresses, which eventually overflow the ESP, suggesting that mcopy or the FAT32-utilities it makes use of aren't aware of the partition's bounds. This commit doesn't yet fix the problem, just includes a report and some scripts that help in showcasing the problem.